feat: add restake btn in staking incentives card
